POC Generate more DBT Models to handle incremental
