+++ .gitignore
@@ -5,6 +5,7 @@ build
 .DS_Store
 data
 .dockerversions
+.vscode
 
 secrets
