+++ airbyte-integrations/bases/base-normalization/normalization/transform_catalog/helper.py
@@ -22,25 +22,6 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
 
-from setuptools import find_packages, setup
 
-setup(
-    name="normalization",
-    description="Normalizes data in the destination.",
-    author="Airbyte",
-    author_email="contact@airbyte.io",
-    url="https://github.com/airbytehq/airbyte",
-    packages=find_packages(),
-    install_requires=["airbyte-protocol", "dbt>=0.18.1", "pyyaml"],
-    package_data={"": ["*.yml", "*.sql"]},
-    setup_requires=["pytest-runner"],
-    entry_points={
-        "console_scripts": [
-            "transform-config=normalization.transform_config.transform:main",
-            "transform-catalog=normalization.transform_catalog.transform:main",
-        ],
-    },
-    extras_require={
-        "tests": ["airbyte-protocol", "pytest"],
-    },
-)
+def jinja_call(command: str) -> str:
+    return "{{ " + command + " }}"
